Restored explicit cast that was removed in 7af5d437e49878ea4e8de73a02d99ad4e5751933.
Found with [-Wsign-compare] while building with MINGW: pointer
difference is of different size on linux and windows.

+++ re2c/src/parse/scanner.cc
@@ -349,7 +349,7 @@ void Scanner::fill (uint32_t need)
 		{
 			need = BSIZE;
 		}
-		if (top - lim < need)
+		if (static_cast<uint32_t> (top - lim) < need)
 		{
 			const size_t copy = static_cast<size_t> (lim - bot);
 			char * buf = new char[copy + need];
